Moved AudioContextManager to pxtlib and updated all calls to it
